+++ tests/AtomUI.Desktop.Controls.Tests/Input/TextBoxVisualStateTests.cs
@@ -1,10 +1,19 @@
 using System;
+using System.IO;
 using System.Linq;
+using System.Reflection;
+using AtomUI;
 using AtomUI.Controls;
+using AtomUI.Controls.Commons;
+using AtomUI.Controls.Primitives;
+using AtomUI.Theme;
+using AtomUI.Desktop.Controls.DesignTokens;
 using AtomUI.Theme.Styling;
 using Avalonia;
+using Avalonia.Animation;
 using Avalonia.Controls;
 using Avalonia.Controls.Presenters;
+using Avalonia.Controls.Primitives;
 using Avalonia.Media;
 using Avalonia.Threading;
 using Avalonia.VisualTree;
@@ -46,6 +55,254 @@ public class TextBoxVisualStateTests
             BrushShouldHaveSameColor(
                 scrollViewer.Foreground,
                 GetThemeResource<IBrush>(SharedTokenKind.ColorTextDisabled));
+        });
+    }
+
+    [Fact]
+    public void TextBox_InnerBoxDecorator_Uses_TextBox_Border_Tokens_For_Default_Hover_And_Focus()
+    {
+        var textBox = new AtomUITextBox
+        {
+            Width           = 180,
+            IsMotionEnabled = false
+        };
+
+        ShowInWindow(textBox, () =>
+        {
+            var border = FindTemplatePart<PixelAlignedBorder>(textBox, "InnerBoxDecorator");
+            var expectedThickness = GetTextBoxTokenResource<Thickness>("BorderThickness");
+            var defaultBorder     = GetTextBoxTokenResource<IBrush>("BorderColor");
+            var hoverBorder       = GetTextBoxTokenResource<IBrush>("HoverBorderColor");
+            var activeBorder      = GetTextBoxTokenResource<IBrush>("ActiveBorderColor");
+            var activeShadow      = GetTextBoxTokenResource<BoxShadows>("ActiveShadow");
+
+            textBox.BorderThickness.ShouldBe(expectedThickness);
+            border.BorderThickness.ShouldBe(expectedThickness);
+            BrushShouldHaveSameColor(border.BorderBrush, defaultBorder);
+
+            SetPseudoClass(textBox, StdPseudoClass.PointerOver, true);
+            Dispatcher.UIThread.RunJobs();
+
+            border.BorderThickness.ShouldBe(expectedThickness);
+            BrushShouldHaveSameColor(border.BorderBrush, hoverBorder);
+
+            textBox.Focus();
+            Dispatcher.UIThread.RunJobs();
+
+            textBox.IsFocused.ShouldBeTrue();
+            border.BorderThickness.ShouldBe(expectedThickness);
+            BrushShouldHaveSameColor(border.BorderBrush, activeBorder);
+            border.BoxShadow.ShouldBe(activeShadow);
+        });
+    }
+
+    [Fact]
+    public void TextBox_Uses_Customizable_SizeType_Contract_And_TextBox_Token_Scope()
+    {
+        typeof(ICustomizableSizeTypeAware).IsAssignableFrom(typeof(AtomUITextBox)).ShouldBeTrue();
+
+        var property = typeof(AtomUITextBox).GetProperty(
+            "SizeType",
+            BindingFlags.Instance | BindingFlags.Public | BindingFlags.FlattenHierarchy);
+        property.ShouldNotBeNull();
+        property!.PropertyType.ShouldBe(typeof(CustomizableSizeType));
+
+        var field = typeof(AtomUITextBox).GetField(
+            "SizeTypeProperty",
+            BindingFlags.Static | BindingFlags.Public | BindingFlags.FlattenHierarchy);
+        field.ShouldNotBeNull();
+        field!.FieldType.ShouldBe(typeof(StyledProperty<CustomizableSizeType>));
+
+        var textBoxTokenType = typeof(AtomUITextBox).Assembly.GetType("AtomUI.Desktop.Controls.TextBoxToken");
+        textBoxTokenType.ShouldNotBeNull();
+        var textBoxScopeProvider = textBoxTokenType!
+            .GetField("ScopeProvider", BindingFlags.Static | BindingFlags.Public | BindingFlags.FlattenHierarchy)
+            .ShouldNotBeNull()
+            .GetValue(null);
+
+        var textBox       = new AtomUITextBox();
+        var scopeProvider = ControlTokenResourceScopeHost.GetTokenResourceScopeProvider(textBox);
+        scopeProvider.ShouldBeSameAs(textBoxScopeProvider);
+    }
+
+    [Fact]
+    public void TextBox_Theme_Uses_TextBox_Token_Resources()
+    {
+        var source = ReadRepoFile("src/AtomUI.Desktop.Controls/Input/Themes/TextBoxTheme.axaml");
+
+        source.ShouldContain("TextBoxTokenResource BorderColor");
+        source.ShouldContain("TextBoxTokenResource BorderThickness");
+        source.ShouldContain("TextBoxTokenResource BorderRadiusLG");
+        source.ShouldContain("TextBoxTokenResource BorderRadius");
+        source.ShouldContain("TextBoxTokenResource BorderRadiusSM");
+        source.ShouldContain("TextBoxTokenResource HoverBorderColor");
+        source.ShouldContain("TextBoxTokenResource ActiveBorderColor");
+        source.ShouldContain("TextBoxTokenResource ActiveShadow");
+        source.ShouldContain("TextBoxTokenResource PaddingLG");
+        source.ShouldContain("TextBoxTokenResource Padding");
+        source.ShouldContain("TextBoxTokenResource PaddingSM");
+        source.ShouldContain("SharedTokenResource UniformlyPaddingXXS");
+        source.ShouldContain("SharedTokenResource FontHeightLG");
+        source.ShouldContain("SharedTokenResource FontHeight");
+        source.ShouldContain("SharedTokenResource FontHeightSM");
+        source.ShouldContain("SharedTokenResource FontSizeLG");
+        source.ShouldContain("SharedTokenResource FontSize");
+        source.ShouldContain("SharedTokenResource FontSizeSM");
+        source.ShouldContain("SharedTokenResource ColorTextPlaceholder");
+        source.ShouldContain("SharedTokenResource ColorTextDisabled");
+        source.ShouldNotContain("LineEditTokenResource");
+        source.ShouldNotContain("AddOnDecoratedBoxTokenResource");
+        source.ShouldNotContain("AddOn");
+        source.ShouldNotContain("TextBoxTokenResource LineHeight");
+        source.ShouldNotContain("TextBoxTokenResource InputFontSize");
+        source.ShouldNotContain("TextBoxTokenResource PlaceholderColor");
+        source.ShouldNotContain("TextBoxTokenResource ContentMargin");
+        source.ShouldNotContain("TextBoxTokenResource DisabledColor");
+        source.ShouldNotContain("TextBoxTokenResource AddOnSpacing");
+        source.ShouldNotContain("InnerLeftContent");
+        source.ShouldNotContain("InnerLeftContentTemplate");
+        source.ShouldNotContain("InnerRightContent");
+        source.ShouldNotContain("InnerRightContentTemplate");
+        source.ShouldNotContain("PART_LeftInnerContentLayout");
+        source.ShouldNotContain("PART_RightInnerContentLayout");
+        source.ShouldNotContain("PART_LeftInnerContent");
+        source.ShouldNotContain("PART_RightInnerContent");
+        source.ShouldContain("<atom:InputClearIconButton");
+        source.ShouldContain("Name=\"PART_ClearButton\"");
+        source.ShouldContain("<atom:RevealButton");
+        source.ShouldContain("Name=\"PART_RevealButton\"");
+    }
+
+    [Fact]
+    public void TextBox_Token_Only_Exposes_TextBox_Visual_Resources()
+    {
+        Enum.GetNames(typeof(TextBoxTokenKind))
+            .OrderBy(name => name)
+            .ShouldBe(new[]
+            {
+                "ActiveBorderColor",
+                "ActiveShadow",
+                "BorderColor",
+                "BorderRadius",
+                "BorderRadiusLG",
+                "BorderRadiusSM",
+                "BorderThickness",
+                "HoverBorderColor",
+                "Padding",
+                "PaddingLG",
+                "PaddingSM"
+            });
+    }
+
+    [Theory]
+    [InlineData(CustomizableSizeType.Large, "PaddingLG")]
+    [InlineData(CustomizableSizeType.Middle, "Padding")]
+    [InlineData(CustomizableSizeType.Small, "PaddingSM")]
+    [InlineData(CustomizableSizeType.Custom, "Padding")]
+    public void TextBox_Padding_Follows_TextBox_Size_Tokens(CustomizableSizeType sizeType, string tokenKind)
+    {
+        var textBox = new AtomUITextBox
+        {
+            Width           = 180,
+            SizeType        = sizeType,
+            IsMotionEnabled = false
+        };
+
+        ShowInWindow(textBox, () =>
+        {
+            textBox.Padding.ShouldBe(GetTextBoxTokenResource<Thickness>(tokenKind));
+        });
+    }
+
+    [Fact]
+    public void TextBox_Default_Motion_State_Follows_EnableMotion_Resource()
+    {
+        var textBox = new AtomUITextBox
+        {
+            Width = 180
+        };
+
+        ShowInWindow(textBox,
+            () =>
+            {
+                var border = FindTemplatePart<PixelAlignedBorder>(textBox, "InnerBoxDecorator");
+
+                textBox.IsMotionEnabled.ShouldBeFalse();
+                border.Transitions.ShouldBeNull();
+            },
+            window => window.Resources[SharedTokenKind.EnableMotion] = false);
+    }
+
+    [Fact]
+    public void TextBox_InnerBoxDecorator_Transitions_Animate_Background_Border_And_Shadow()
+    {
+        var textBox = new AtomUITextBox
+        {
+            Width           = 180,
+            IsMotionEnabled = true
+        };
+
+        ShowInWindow(textBox, () =>
+        {
+            var border = FindTemplatePart<PixelAlignedBorder>(textBox, "InnerBoxDecorator");
+            var transitionProperties = border.Transitions.ShouldNotBeNull()
+                                             .Select(transition => transition.Property)
+                                             .ToArray();
+
+            transitionProperties.ShouldContain(PixelAlignedBorder.BackgroundProperty);
+            transitionProperties.ShouldContain(PixelAlignedBorder.BorderBrushProperty);
+            transitionProperties.ShouldContain(PixelAlignedBorder.BoxShadowProperty);
+        });
+    }
+
+    [Fact]
+    public void TextBox_Clear_And_Reveal_Buttons_Use_LineEdit_Default_Icons()
+    {
+        var textBox = new AtomUITextBox
+        {
+            Width                = 180,
+            Text                 = "secret",
+            PasswordChar         = '*',
+            IsAllowClear         = true,
+            IsEnableRevealButton = true,
+            IsMotionEnabled      = false
+        };
+        var lineEdit = new AtomUILineEdit
+        {
+            Width                = 180,
+            Text                 = "secret",
+            PasswordChar         = '*',
+            IsAllowClear         = true,
+            IsEnableRevealButton = true,
+            IsMotionEnabled      = false
+        };
+        var layout = new StackPanel
+        {
+            Children =
+            {
+                textBox,
+                lineEdit
+            }
+        };
+
+        ShowInWindow(layout, () =>
+        {
+            var textBoxClear  = FindTemplatePart<InputClearIconButton>(textBox, "PART_ClearButton");
+            var lineEditClear = FindTemplatePart<InputClearIconButton>(lineEdit, "PART_ClearButton");
+            var textBoxReveal  = FindTemplatePart<RevealButton>(textBox, "PART_RevealButton");
+            var lineEditReveal = FindTemplatePart<RevealButton>(lineEdit, "PART_RevealButton");
+
+            textBoxClear.IsVisible.ShouldBeTrue();
+            lineEditClear.IsVisible.ShouldBeTrue();
+            textBoxReveal.IsVisible.ShouldBeTrue();
+            lineEditReveal.IsVisible.ShouldBeTrue();
+
+            textBoxClear.Icon.ShouldNotBeNull();
+            lineEditClear.Icon.ShouldNotBeNull();
+            textBoxClear.Icon!.GetType().ShouldBe(lineEditClear.Icon!.GetType());
+
+            AssertSameIconType(textBoxReveal, lineEditReveal, AbstractToggleIconButton.CheckedIconProperty);
+            AssertSameIconType(textBoxReveal, lineEditReveal, AbstractToggleIconButton.UnCheckedIconProperty);
         });
     }
 
@@ -180,6 +437,12 @@ public class TextBoxVisualStateTests
         return (T)value!;
     }
 
+    private static T GetTextBoxTokenResource<T>(string kindName)
+    {
+        var tokenKind = (TextBoxTokenKind)Enum.Parse(typeof(TextBoxTokenKind), kindName);
+        return GetThemeResource<T>(tokenKind);
+    }
+
     private static void BrushShouldHaveSameColor(IBrush? actual, IBrush? expected)
     {
         GetSolidBrushColor(actual).ShouldBe(GetSolidBrushColor(expected));
@@ -202,7 +465,24 @@ public class TextBoxVisualStateTests
         return ((ISolidColorBrush)brush!).Color;
     }
 
-    private static void ShowInWindow(Control content, Action assertion)
+    private static void SetPseudoClass(Control control, string pseudoClass, bool value)
+    {
+        ((IPseudoClasses)control.Classes).Set(pseudoClass, value);
+    }
+
+    private static void AssertSameIconType(AvaloniaObject actualOwner,
+                                           AvaloniaObject expectedOwner,
+                                           StyledProperty<PathIcon?> property)
+    {
+        var actual   = actualOwner.GetValue(property);
+        var expected = expectedOwner.GetValue(property);
+
+        actual.ShouldNotBeNull();
+        expected.ShouldNotBeNull();
+        actual!.GetType().ShouldBe(expected!.GetType());
+    }
+
+    private static void ShowInWindow(Control content, Action assertion, Action<AvaloniaWindow>? configureWindow = null)
     {
         var window = new AvaloniaWindow
         {
@@ -210,6 +490,7 @@ public class TextBoxVisualStateTests
             Height  = 120,
             Content = content
         };
+        configureWindow?.Invoke(window);
 
         try
         {
@@ -221,5 +502,29 @@ public class TextBoxVisualStateTests
         {
             window.Close();
         }
+    }
+
+    private static string ReadRepoFile(string relativePath)
+    {
+        var path = GetRepoFile(relativePath);
+        File.Exists(path).ShouldBeTrue($"Expected repository file to exist: {relativePath}");
+        return File.ReadAllText(path);
+    }
+
+    private static string GetRepoFile(string relativePath)
+    {
+        var directory = new DirectoryInfo(AppContext.BaseDirectory);
+        while (directory is not null)
+        {
+            var candidate = Path.Combine(directory.FullName, relativePath);
+            if (File.Exists(candidate))
+            {
+                return candidate;
+            }
+
+            directory = directory.Parent;
+        }
+
+        return Path.Combine(AppContext.BaseDirectory, relativePath);
     }
 }